Add Memcache support as a storage for rate limiting
In order to implement the memcache support the rate limiting method is
slightly changed. Memcache does not have a method for counting all keys
which start with a certain prefix. Such a method was used in the Redis
implementation.

Because of that the concept of "bucket" is introduced. A bucket is an
interval of time equal to the set rate limiting interval. All requests
for particular client are counted in a bucket by incrementing its value.

This approach has a huge performance gain since only one key-value would
be stored for particular client. This is in contrast with `$maxRequests`
key-value pairs for the previous approach.

Unfortunately the current implementation introduces a race condition. It
is possible for parallel requests to read the same value for the client
and all of the update it with `oldValue + 1`. Thus one can actually
exceed `$maxReqeusts`. Because the rate limiter is not meant to be super
strict I think this behaviour is fine in the end.